Version 7
We have implemented website adaptation specifically for screens wider than 1440px to ensure an optimal viewing experience for users with large monitors. Additionally, we are pleased to introduce a Ukrainian version of the site to better serve our audience. We have also fixed all known issues in the footer, enhancing its functionality and visual appeal.